Extract attachment rendering into partial
Previously publication and consultation attachments were rendered
inline. This extracts the rendering into a new 'attachments' partial,
which has the nice side effect of improving the aria markup for
consultations, as well as making the testing approach consistent.

This also replaces the 'documents' question methods in each presenter
with explicit calls in the views. Since the view needs to be aware of
the type of data, it makes sense to ask questions like 'present?' and
'any?' in the view directly, instead of bloating the presenters.