Fix usages of aria-labelledby
Previously the div of attachments used 'aria-labelledby' to
refer to an outer heading element, which is incorrect: the
label should refer to something inside the 'div'. The 'div'
should also set the appropriate 'role="section"' attribute [1],
or we can just replace the 'div' with a 'section' element [2].

This corrects the structure of the attachments partial so
that the sections properly self-describe. Each section now has
an ID, so they continue to be easy to grab hold of in tests.

[1] https://www.w3.org/TR/wai-aria/#region

[2] https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/Region_role