Use presenter helps across all presenter tests
This adds a new method in, create_presenter, which has default test
arguments set for presenters. This then updates all of the presenter
tests so that they all use this helper.

The reason for this change is that I intend to add a third argument soon to
presenter classes which would mean an increase in test verbosity. I also
think this improves consistency of these tests.