Use requested_path rather than requested_content_item_path
This shortens this variable name and I believe it still signifies the
same intent as the previous name. The optional nature of it is also
removed since every time this class is initialised in app this attribute
is set, I imagine this was set as optional for test verbosity but this
is no longer an issue.

This also fixes a couple of ivars that weren't necessary due to methods
existing.