Noindex withdrawn content for crawlers
We publish a sitemap with down-weighting for withdrawn content, however this
does not seem to affect the ranking of external search results as much as we
might hope.

By no-indexing here, we hope to surface current information more readily in
places like google.

Should users wish to find withdrawn content, then it'll be findable via site
search or navigation.