Use Plek#asset_root over Plek#public_asset_host
I'm removing this as part of a plan to remove Plek#public_asset_host and
the associated GOVUK_ASSET_HOST env var. The reason to remove these is
because they are the same #asset_root and GOVUK_ASSET_ROOT, which pushes
developers to consider which one to use when they both have the same
effect.

Government frontend appears to be the only app other than Whitehall that
has used this method, which I assume is because the code it is used in
was ported from Whitehall.