Remove jasmine-rails gem
According to searls/jasmine-rails#238 the jasmine-rails gem is no
longer actively maintained and is reliant on the phantomjs headless
browser which has suspended development (ariya/phantomjs#15344).
Therefore it makes more sense to switch to the jasmine gem and use the
jasmine_selenium_runner gem to switch to headless Chrome for testing.