Remove gem "asset_bom_removal-rails"
To work around the issue with sassc and unquote the accepted solution has been to
set css_compressor to nil and then use sass_style to achieve a similar effect.
However the gem forces the css_compressor back to :sass thus breaking again on asset compilation.

Removing the gem fixes this and the gem itself does not seem to be necessary since
it was introduced to fix issues in Firefox 52 and the latest Firefox version is
78