Test alternative text for Transition navigation
We want to test whether changing the text on the banner navigation improves
click through rate. We want to keep the test simple, so we've just
chosen a [single, high traffic page](https://www.gov.uk/visit-europe-1-january-2021) as
victim.

The displayed navigation is determined in govuk_publishing_components, however
we know how it works. If a certain taxonomy content_id is found in the taxon links
of a content item, then a step by step banner (colloquially a "super breadcrumb")
will be displayed using the title of the taxon.

If we intercept the content item before it is passed to the component, then
we can tweak what is displayed.

https://trello.com/c/igPWXPvz/482-put-new-transition-taxon-live