Swap sanitize for raw in detail guide template
The `sanitize` method was stripping out tables for reasons. This replaces it
with the slightly more risky `raw` method that doesn't break the markup. Using
`sanitize`s allow list meant listing most of the HTML elements, and introducing
a risk of further breakages if a element is accidently left out of the
allow list.