Don't sort facets by type
This behaviour is slightly confusing because it re-orders the facets
that are already in a definied order in Specialist Publisher. This
means that content designers might be expecting the order in a
certain way and then when it gets to the frontend the ordering is
different.

A very clear example of this is if you had three facets, one `date`, one
`text` and one `time`. In the finder the facets might be ordered such
that you have the date, the time and then the text field.

However this code will re-order that so you end up with the date and
time field separated by the text field in between them.

This change was added in 4f0f5a63b9127717bc738ac1375ef80f853979f5 but I
would argue that a better solution is to avoid having this sort of logic
in the frontend app and instead make it such that the publishing app
decides the ordering of the fields. This means we can ensure that the
fields in Specialist Publisher show up in the same order as on the
frontend. We've recently been dealing with a department and there was
some confusion over this because the order they were seeing in
Specialist Publisher (which was the order they wanted) was not the order
they saw when previewing their documents.

I've checked the other Specialist Finders and we won't negatively affect
any of the existing ones, the dates will always be grouped together,
but some of them may now appear at the bottom of the blue box instead of
the top. But then again, this is now consistent with the ordering of the
facets in the finder itself.