Modify Get Involved page and style for new framework
When Get Involved existed in Whitehall it relied on an outdated set of
styling frameworks that we no longer support. Of course, none of these
were present in Government Frontend, so a large chunk of the related
work was in modernising this style system.

Also included in the Get Involved page specifically are calls to the
correct functions in the controller such that we have the needed
data to display.