Remove assumptions of attachment existence
This resolves a couple of scenarios where assumptions are made that
documents will have attachments and attachments with particular ids
exist. This is data that developers assume is the case (and most likely
is) but is beyond the guarantees that schemas make.

This mismatch can cause tests to fail when valid schemas are generated
which cannot be rendered.