Update back-to-top SVG arrow attributes
It was recommended on GOVUK Accounts team's DAC report that
focusable="false" and aria-hidden="true" are added to the "up arrow" SVG
on our Privacy notice and Accessibility statement page.
This is because SVGs are in tab sequence in IE, which means that focus
disappears as it sequences over them. Adding focusable="false" ensures
the SVG is not in tab sequence, while aria-hidden="true" ensures that it
is hidden from screen readers.