Remove Downing Street Briefings code
We (GOV.UK) added custom functionality to the stack in order to support bespoke
needs for the Downing St. briefings page; Number 10 since cancelled
the briefings removing the need for the page; Thus there is now code
on GOV.UK that will never be executed. This commit removes said code.