Remove the dependencies section
This is too vague to be actionable, and Content Store / Static are
common dependencies for frontend apps that we can document centrally.
I've kept the reference to ImageMagick, as it links in with existing
docs, but it's unclear why anyone needs to know about PhantomJS in
order to work on this repo.