Remove faulty screenshot guidance
This doesn't work with GOV.UK Docker since the screenshots are saved
to the "tmp" directory, which is a temporary volume. We could change
the guidance to advise specifying a path.

On the other hand, a README isn't the right place to be documenting
features of a third-party library, which aren't specific to this repo
or GOV.UK. We can always write a dev doc if necessary.