Do not check if a page is saved if a user has no session header
I spotted that the account-api was getting a *lot* of saved-page
queries in staging, which I tracked down to this: we're checking if
the page is saved even if the user has no session (which will always
fail).  We can skip the API call in that case.

We had a similar issue once in the Transition Checker, where we were
trying to fetch saved results regardless of whether the user had a
session or not.  Since this has come up twice now we should think
about how this problem can be solved generically.  Maybe we could put
more logic in gds-api-adapters, for example?