Migrate get-involved data sourcing to Search API
Previously we were retrieving all data for the get involved page from
the Publishing API. This was particularly heavy on our system with
extra unneeded pings that were not correctly cached. Moving to the
Search API where possible allows us to use cached results and saves
a lot of overhead.