remove jq from ga tracker and application.js
The only code in application.js is to initialise the ga tracker
so they're being done at the same time. I've removed one `get`
but not the other because the `tracker` object is defined by google
while the `form` object is a standard HTML element

You will notice that the application.js calls on the `getElementsByClassName`
method which can return multiple results. Then we're only using the first
element of that list. This isn't ideal but we have confidence that this
class can only be used once on each page which makes this fine.

Call route for the set-ga-client-id-on-form
- application.js is the only file that calls on it
- it only applies it to elements with the `js-service-sign-in-form` class
- that class is only created by [this partial](https://github.com/alphagov/government-frontend/blob/main/app/views/content_items/service_sign_in/_choose_sign_in.html.erb#L25)
- that partial is only rendered by [this page](https://github.com/alphagov/government-frontend/blob/main/app/views/content_items/service_sign_in.html.erb#L8) by way of a method in [this presenter](https://github.com/alphagov/government-frontend/blob/main/app/presenters/service_sign_in/choose_sign_in_presenter.rb#L5)
- This means we can be confident there's only one element of that class on a page, huzah!