Replace heading with title
When defining the main page header we use the title component by convention. This renders an H1 element with the associated styles. We also remove the elements that wrap this component as none of classes used are defined anywhere in this application (`block`, `headings-block`, `inner-block` or `floated-children`) – presumably remnants from when moved from whitehall.