Fix in-page link to #take-part
There are two main headings and associated sections on this page `engage-with-government` and `take-part`, both linked to from the lead paragraph. The `take-part` section was given the wrong id, presumably by copying the previous section, so we correct this. Because of the broken layout of the page (a set of floating elements) these sections need a clear property, otherwise the in-page link won't work.