Remove jQuery from track-radio-group.js (except the Ajax request)
This change removes - most - of the jQuery from the track-radio-group.js
module and replaces it with vanilla JS. It doesn't replace the '$.ajax()'
function call - that will happen in a separate commit.

It was noted when replacing the 'track' event handler for 'submit' that the
tests became broken. This was put down to jQuery handling 'trigger' differently
now that it is a DOM element. A change to triggering 'click' on the forms
submit button resolved the issue, but caused an error where the form was no
longer attached to the document. Appending it in beforeEach() and removing it in
afterEach() resolved the issue.