Include account session flash in Vary header
When a user subscribes or unsubscribes to a single page email
notification topic, they'll be redirected back to the content page with
a message in their account session flash. This message will be detected
and used to display the relevant success message on the page.

We need to set the vary header to indicate to Fastly that the success
messages are different variations and should be cached and served
separately. The `GovukPersonalisation::ControllerConcern` handles this
for us but sets the header to vary on the full account session. This
would result in pages being cached per-user which isn't very helpful in
this case.

Instead, override the `set_account_vary_header` method to vary on the
flash message instead. The message content won't vary per user so this
will mean we only cache three versions of the page - one without a flash
message, one with the subscribed message and one with the unsubscribed
message.