Add email subscribe/unsubscribe flash messages
When a user completes the single page email subscribe or unsubscribe
journey they will be redirected to the page they were on with a message
in the account session flash.

Check for these messages in the flash and render the appropriate success
banner. The pages chosen for  phase 1 and phase 2 of the rollout are
either `publication` or `detailed_guide` content types, so only add the
partial to these templates. In the future when we roll out to more pages
we can add the partial to the rest of the content types.