Remove jQuery Ajax request from track-radio-group.js
The jQuery '$.ajax()' method used is a [JSON-P (JSON with Padding)][1] request.
It is not however, an Ajax request as it [does not use the 'XMLHttpRequest' object][2].

This is indicated by the use of the 'dataType: 'jsonp'' config option when
calling the jQuery function e.g. '$.ajax({dataType: 'jsonp'})'.

To achieve this ["jQuery attaches a global function to the window object that is
called when the script is injected, then the function is removed on completion."][3].

Essentially, jQuery creates a '<script>' element with a 'src' attribute set to
the URL of the request. A 'callback' function is also created and it's name is
appended to the URL.

When the request is made successfully - the specified 'callback' function is
called with the data returned from the request as the first (and only)
parameter. The '<script>' element is then removed.

This technique allows JavaScript to send data to-and-from foreign origins which
would otherwise fail due to a [CORS (Cross-Origin Resource Sharing)][4] policy.

We need to mimic this behaviour in vanilla JavaScript - and within a module.

Firstly, we add an event trigger that is attached to the 'window.GOVUK' context.
This is our 'callback' function and will signal the completion of the JSON-P
request.

Next we create an event listener for the above trigger which will get the
response data from the event and call the appropriate function inside the
module.

Finally, the jQuery '$.ajax()' function can be replaced by a function that now
creates the '<script>' element and assigns a suitable 'src' - complete with
'callback' name. A 'timestamp' is used to avoid caching.

[1]: https://en.wikipedia.org/wiki/JSONP
[2]: https://www.w3schools.com/js/js_json_jsonp.asp
[3]: https://remysharp.com/2007/10/08/what-is-jsonp/#how-it-works-in-jquery
[4]: https://en.wikipedia.org/wiki/Cross-origin_resource_sharing