Add notification button to publisher metadata
Add the single page notification button to the version of the publisher
metadata component that includes a "See all updates" link (so, when
link_to_history is true)

The notification button requires a base_path therefore a base_path must
be passed to the publisher metadata component where we want the button
to appear.

In the initial stages we do not want the button to appear everywhere
(i.e on every single page that uses a certain type of template or
publishing format) so there is an additional include_notification_button
flag to fine-tune where the button should appear.