Convert sticky-element.js to use the Module pattern
What

Currently, the sticky-element in Government Frontend does not follow the
JavaScript module pattern.

Convert the existing code to be consistent with this pattern.

Why?

The module pattern is the standard approach and this JavaScript component
does not follow the convention.

A note on the tests

The tests for sticky-element do not actually test physical scrolling on
the page. Instead they test for presence of specific CSS classes. Because
of this it was found necessary to run the checkResize() and checkScroll()
methods prior to expecting these classes to have changed. In effect this
simulates physically scrolling or resizing the page.