Support page notification button on consultation
Add support for the single page notification buton to the consultation
template. The button will not appear until the feature is enabled on
pages with this content type.

Seeing as the button is always grouped together with the published dates
component I created a new  `published_dates_with_notification_button`
partial to be shared between publications, detailed_guides and
consultations.