Move single page notification button
The single page notification button component was originally created in
govuk_publishing_components as the intial understanding was that it
might be used on different frontend apps.

However, the button is only being used on government-frontend currently
and it is highly unlikely that there will be a purpose for this
component outside of the scope of this application.

This commit copies over the single_page_notification_button from
govuk_publishing_components into government-frontend's app-level
component library. Some notable adjustments have been made to the test
file (since government-frontend uses Minitest and
govuk_publishing_components relies on Rspec) and also to the single page
notification helper as instance variables are not allowed in helpers in
government-frontend.

The component's JS test file relies on Jasmine Ajax so support for that
has also been added as part of this work.