Implement notification button exemption list
Change single page notification button allow list to an exemption list.

Add tests to verify that pages whose content item is on the exemption
list do not display with the single page notification button.

Ensure that the button appears only on detailed_guides, publications and
consultations and update tests for all other types of publications to
check that the button is not included on those types.