Put account flash message on consultation
This was missed in the previous PR which added the button to the
consultation template. The success message is necessary to indicate when
subscribing/unsubscribing to notifications has been completed

Update test to verify that the success banner appears as expected on all
publication types on which the single page notification button is
present.