Remove Wraith dependencies and configuration
The wraith tests for this repo are no longer a part of the review/CI
process and no longer run. The Wraith gem itself has not had further
development since 2019.

My impetus to remove this was that one of Wraith's dependencies,
chromedriver-helper, actually causes problems by installing a
chromedriver binary that can conflict with system installed ones.