Add attachment arguments
These are needed to render the new accessible form link from GOVUK
Publishing Components. This is dependant on the component being updated
first [1].

This change will apply to publications and consultations, which use the
attachments partial. Initially, the new flow of requesting accessible
formats will be a pilot and only available for documents, owned by a few
organisations. This code will not need to be changed when the change is
released to all orgs.

[1]: https://github.com/alphagov/govuk_publishing_components/pull/2636