Remove jQuery ready function
The "ready" function in jQuery waits for the DOM tree to load, i.e. be
"ready" before moving on. The equivalent to this in vanilla JS is to
listen for the "DOMContentLoaded" event. However, this script works
without an event listener on "DOMContentLoaded", so the ready function
has just been removed.