Remove ajax from library/webchat.js
Creates an XMLHttpRequest object to make the request to the availability
url. We need to use XMLHttpRequest rather than the new "await" and
"fetch" functions as they are not supported in Internet Explorer.

There is only one webchat page that this script runs on:
/government/organisations/hm-passport-office/contact/hm-passport-office-webchat

The [availabilityURL] always returns a response of:

```
{
  "status": "success",
  "response": "UNAVAILABLE"
}
```

with an http status of 200. This seems incorrect as the [openUrl] states
that all advisors are busy, so we would expect the availabilityURL to
return a status of "BUSY". There is an [open ticket] to change the URLs,
so hopefully that will fix the problem. The removal of jQuery in this
change has not negatively affected the workings of Webchat. The
behaviour hasn't been affected, even if that behaviour is incorrect.

The XMLHttpRequest is still being called asynchronously like the old
ajax call ("true" in the call to request.open).

`readyState` and the `status` returned by the XMLHttpRequest object
are used to determine whether the request has been successful. That
means how the response is being stubbed in the tests has been changed to
match.

The "mount" function is doing the equivalent of the webchat.js script,
so it has been updated to match that script.

The jsonNormalisedError object had to be updated in the tests, as a
string without a status  was not an accepted response for
XMLHttpRequest.

As ajax has been removed, it can no longer be spied on in the tests.
However, the tests still use jQuery, so jasmine Ajax (installed in a
prior commit) can be used define fake responses from XMLHttpRequest.

XMLHttpRequest makes two calls, to open and send, it was easier to stub
the response of the most recent call rather than trying to mock both.
This means that in test "should only track once per state change" that
tests the URL polling, we only need to change the "most recent" ajax
stub when the response being returned changes. It was much simpler to
just call the "most recent" stub twice than try to replicate the
operation of the fake function.

[availabilityURL]: https://hmpowebchat.klick2contact.com/v03/providers/HMPO/api/availability.php
[openUrl]: https://hmpowebchat.klick2contact.com/v03/launcherV3.php?p=HMPO&d=717&ch=CH&psk=chat_a1&iid=STC&srbp=0&fcl=0&r=Chat&s=https://hmpowebchat.klick2contact.com/v03&u=&wo=&uh=&pid=2&iif=0
[open ticket]: https://github.com/alphagov/government-frontend/pull/2358