Add shared ManualSection module
**Note**: Test file for this module is being worked on seperately.

This will included for both HMRC manual sections and normal manuals.

Currently, we have to send extra requests to Content Store for data
included in the parent manual for a section page. A better solution
would be to include this data in the section at the point of publishing.

We need to:

* Add the same tagging to a section as a manual for contextual
breadcrumbs (e.g `topics`).

* Add the manual title to the HMRC section content item, as it only
includes the base path.