Uses manual for breadcrumbs over section
Currently (in Manuals Frontend), the contextual breadcrumbs for section
pages are generated via the parent manual content item. We have to do
the same in Government Frontend as the tagging between the section and
manual is different and so if we just used the section content item then
we lose breadcrumbs.