Use static templates for manual updates pages
Update's pages don't exist as a content item in Content Store. The
content item that is returned for the update's pages
(/guidance/<manual>/updates or /hmrc-internal-manuals/<manual>/updates)
is the parent manual content item and so we have to specify the
templates and presenter directly, otherwise Government Frontend will try
and use the parent content item's view and presenter.

I've not added tests for these as the integration tests we're adding,
and the ones that already present cover the new logic.