Replace the controller tests for GetInvolved
This changes the previous tests, which tested a number of controller
methods, into a set of tests that take the more idiomatic Rails approach
of testing the controllers via requests.

The motivation for this work was to enable the retirement of a large
set of tests for this set of pages done through Smokey [1]. These Smokey
tests are unconventional as they seem to be used as a the primary means
to test these pages - this is not a reliable or scalable approach given
they're in a separate repository. Nor is it necessary as the logic is
within this application.

[1]: https://github.com/alphagov/smokey/blob/7553a432259d785fb6d4fb81cfd4fe3572185ad3/features/apps/government_frontend.feature#L23-L140