Remove superflous instance variable from controller
- Rather than set an environment variable in the controller which is available to all
the views, we should control whether or not we show the button template in the
presenter.
- I've removed the ALLOWED_DOCUMENT_TYPES because if we only include the
ContentItem::SinglePageNotificationButton module in the presenter classes of content
types which we want to show the button, the effect is the same. It's also a bit misleading
having an allow list because it infers there's a reason some types can't support email
subscriptions. This is not the case - email alert API doesn't care.

- This pattern of including a module into the document type specific presenter
is more consistent with this repo, and other frontend repos.