Remove rack_strip_client_ip middleware
This removes middleware that strips the Client-IP header from requests to
prevent Rails from raising a "IpSpoofAttackError" exception. Removing
this because it isn't needed, we are stripping the Client-IP header at
the CDN level which should prevent this expection being raised anyway.
If this header is being set else were in the request path, we should fix that
instead of blindly stripping the header altogether.

There is also a bug in the middleware, where it prevent previous
middleware from receiving information from following middleware in the
'env' variable. This is because it creates a copy of env, which isn't
returned to the calling middleware. This prevents middleware such as the
prometheus exporter from functioning properly.