Replace inline styles with override classes
The usage of inline styles is going to be disallowed at the CSP level
to reduce the damage HTML injection can do [1].

This replaces the usages of these inline styles with Design System
override classes. I went for these rather than creating a new class as
there didn't seem to be any precedent for specific classes for this part
of the site, yet a variety of precedence in using override classes.

[1]: https://github.com/alphagov/govuk_app_config/pull/279