Renable test/contracts/govuk_content_schemas_test.rb
There was a typo in the support file, such that
GovukSchemas::Example.find_all(supported_schemas) was returning []
because the `find_all` method expects a single string param.
https://github.com/alphagov/govuk_schemas/blob/4fef0aa95be24c3534a00b622a59ea7f1e8175fa/lib/govuk_schemas/example.rb#L5

Also removes redirect, as there is no frontend rendering.
https://github.com/alphagov/publishing-api/tree/main/content_schemas/examples/redirect

Minitest won't not run tests of the same name, and multiple schema examples have
the same basepath. Hence adding in the index to the iterator.