Add `how_government_works` as a singular inflection
We use the `classify` method to work out the presenter name from the
schema name [1]. By default, Rails singularizes names when turning them
into a class.

This led to us incorrectly looking for the `HowGovernmentWorkPresenter`.

Adding an inflection to tell Rails that this string is the same where
singlular or plural, therefore leading to the correct presenter name
being used.

1: https://github.com/alphagov/government-frontend/blob/e0c26e4d7685994e1def3a19751acc7e49c97ab5/app/services/presenter_builder.rb#L41