Remove banner from DCMS about page
Content request: DCMS [1] has now been closed and we can remove the org is
changing banner. department-for-business-energy-and-industrial-strategy
and department-for-international-trade will be closing soon, at which
point we can remove this bespoke banner code entirely.

We also need to remove the banner code for the actual org page which
lives in collections. Once an org is closed in Whitehall the bespoke
banner gets replaced automatically as a different view is rendered [2]
vs [3]. This means we don't have to remove each orgs custom banner as it
gets replaced anyway. We will still need to remove the banner code in
Collections when all three orgs are closed.

[1]: https://www.gov.uk/government/organisations/department-for-digital-culture-media-sport/
[2]: https://github.com/alphagov/collections/blob/6ded61a31a81b7c053151bf74c80c2cd84d001e8/app/views/organisations/separate_website.html.erb#L11-L13
[3]: https://github.com/alphagov/collections/blob/6ded61a31a81b7c053151bf74c80c2cd84d001e8/app/views/organisations/show.html.erb#L7-L13