Quick dartsass-rails proof of concept
This is a quickly put together integration of this app with dart-sass to
experiment with migrating away from sassc-rails.

At a basic level this works for the most part (but may lack some
images/fonts) however it is incomplete with pain points to resolve.

These are:

1) Integration with gouvk_publishing_components

One of the changes of using dartsass-rails with sprockets is that manifest.js
files can no longer reference css files (they're instead compiled to an
assets/build directory). This is a challenge because
govuk_publishing_components' manifest.js has a lot of references to
these files [1].

To resolve this we would perhaps want to consider having
govuk_publishing_components be configured for dartsass usage and have a
separate manifest used for dartsass instances without css files that
configures dartsass with which files to include

2) A dev experience with CSS automatically rebuilt

With dartsass-rails CSS is compiled with a separate binary to the Rails
server that sprockets runs on. Thus to run the Rails app and have
automatic CSS compiling two processes will need to be run. dartsass
rails conveniently provides a procfile.dev so that these can be run by
foreman. However to use this we'd need to update our startup.sh and
govuk-docker configurations to use foreman.

[1]: https://github.com/alphagov/govuk_publishing_components/blob/19bd3d00a13484b25a62895353c128891b379d18/app/assets/config/govuk_publishing_components_manifest.js