Makes travel advice agnostic to presence of 'details -> summary' in its data
These changes are part of a piece of work to move the 'summary' section in travel advice from 'details -> summary' to 'details -> parts'. As such, the government-frontend needs to render as expected whether 'details -> summary' is included in the data for a given country, or whether it should just render the first part for the first page instead. After we have migrated the travel advice data to move the summary content from 'details -> summary' to 'detail -> parts', we will return to remove the agnostic functionality and tests that account for the old data schema.