Remove code for Coming Soon document type
Whitehall stopped making use of this document type in 2018 [1] and there
are no longer any live ones. This removes the code for this document
type from this app.

The motivation for making this change was to remove the schema and
document type from Publishing API [2] which requires code removal from
here for the schema tests to pass.

[1]: https://github.com/alphagov/whitehall/pull/4851
[2]: https://github.com/alphagov/publishing-api/pull/2369