Remove redundant travel advice tests
This is the first part of the cleanup following moving summary into parts for travel advice items. This commit removes tests that are no longer relevant as well as temporarily commenting out the travel advice presenter tests. The latter enables us to remove the optional summary from the travel advice schema in publishing api before we come back to tweak these tests and remove the logic to check for summary as part of details as summary will now always be a part.