Add JavaScript to hide extra world location links
The "hide other links" JavaScript from Whitehall is used to hide long lists of
world locations in the header of the Worldwide Office and Worldwide
Organisation pages.

This moves it across to government-frontend, removing any jQuery, unneeded
features and configurations options. I have rewritten the module in the style
of the other JavaScript in this application.